// src/app/models/user.ts
export class User {
    id: number;
    first_name: string;
    last_name: string;
    email: string;
    username: string;
    job_title: string;
    company: string;
    employees: string;
    phone_number: string;
    country: string;
    password: string;
    rol: string;
    enabled: string;

}